Code refactoring
Hit a problem with the elements of my wxTextCtrl array not getting added in the correct position to the sizer. If I can't fix it I will look at a different method of refactoring the code
