+++ Capstone Code/Project2/Debug/Project2.log
@@ -1,9 +1,8 @@
 ﻿E:\VisualStudioEnterprise2017\Common7\IDE\VC\VCTargets\Microsoft.CppBuild.targets(387,5): warning MSB8028: The intermediate directory (Debug\) contains files shared from another project (Project2.vcxproj).  This can lead to incorrect clean and rebuild behavior.
-  Sim.cpp
-c:\users\tapoz\documents\github\inferno\capstone code\project2\sim.cpp(128): warning C4244: 'initializing': conversion from 'double' to 'int', possible loss of data
-c:\users\tapoz\documents\github\inferno\capstone code\project2\sim.cpp(131): warning C4244: '=': conversion from 'float' to 'int', possible loss of data
-c:\users\tapoz\documents\github\inferno\capstone code\project2\sim.cpp(147): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
-c:\users\tapoz\documents\github\inferno\capstone code\project2\sim.cpp(148): warning C4018: '<': signed/unsigned mismatch
-c:\users\tapoz\documents\github\inferno\capstone code\project2\sim.cpp(149): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
+  Source.cpp
+c:\users\tapoz\documents\github\inferno\capstone code\project2\source.h(37): warning C4091: 'static ': ignored on left of 'const displayType' when no variable is declared
+c:\users\tapoz\documents\github\inferno\capstone code\project2\source.cpp(102): warning C4018: '<': signed/unsigned mismatch
+c:\users\tapoz\documents\github\inferno\capstone code\project2\source.cpp(430): warning C4129: ':': unrecognized character escape sequence
+c:\users\tapoz\documents\github\inferno\capstone code\project2\source.cpp(430): warning C4129: ' ': unrecognized character escape sequence
   Project2.vcxproj -> C:\Users\tapoz\Documents\GitHub\Inferno\Capstone Code\Debug\CapstoneBase.exe
   Project2.vcxproj -> C:\Users\tapoz\Documents\GitHub\Inferno\Capstone Code\Debug\CapstoneBase.pdb (Partial PDB)
