+++ Capstone Code/Project2/Project2.vcxproj
@@ -152,6 +152,7 @@
     <ClInclude Include="mathplot.h" />
     <ClInclude Include="Sim.h" />
     <ClInclude Include="SimIDs.h" />
+    <ClInclude Include="Source.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
